requirements.txt
chardet==3.0.3
memory-profiler==0.47
pdfminer.six==20170419
pdfminer3k==1.3.1
ply==3.10
py==1.4.33
pycrypto==2.6.1
pytest==3.1.1
six==1.10.0
